+++ backend/payment_processing/src/main/java/com/devsquad/payment_processing/service/AccountService.java
@@ -19,7 +19,7 @@ public class AccountService {
         return accountRepository.getAllAccountsR();
     }
 
-    public Account getAccountByIdS(Integer accountNumber) {
+    public Account getAccountByIdS(Long accountNumber) {
         return accountRepository.getAccountByIdR(accountNumber);
     }
 }
